Propose to split 2.1 milestone
Currently there is a long queue of development items in the 2.0 milestone. Because master is unstable and under heavy changes, no one other than @NGPixel can contribute to open source development.

I would like to suggest that we do all tickets which are expected to cause heavy changes first, then do "cosmetic" or "add-on" type features which can be implemented without major changes to the core in the 2.1 milestone while allowing 2.0 to stabilize. This will be hopefully allow the community to contribute some of the 2.1 patches so we all get what we want sooner.