ENH: passthrough to blob content in the repository
There may be relative links to other files on the wiki page. The most
common situation - images. It's much conveniently to store these images
near the md file and not in the upload directory (when you manually
commits to the repo). Wikijs should recognize such requests and
passthrough them, returning image. It shouln'd treat them as wiki-page query.