misc: Match sample git URL to auth type
When stepping through graphical configuration, the sample git URL is
https while the sample type is ssh. This leads to confusion, as it
appears that the https URL should be inserted into the box even when ssh
authentication is used. While the input placeholder could clear up this
confusion, it is hidden by the default value. Change the default value
to match the auth type.