Fix list item formatting for lists with paragraphs
This removes the fancy tick-mark next to unordered list items, which is
a little depressing, but fixes formatting for them when they contain
paragraphs (`<p>` elements). The tick-mark has been replaced with
a square mark. Both ordered and unordered lists have grown 1rem of left
padding to compensate for their position.

Considered using disclosure-closed as a similar triangle replacement,
but it's not regarded as something that should be used in production
(per MDN). Using a simple string is unsupported because it's only in
a CSS working draft, so a square is the least controversial option,
I think.

Closes #1282.