add env variables for custom certificate usage
This would make it possible to use a custom certificate and therefore we gain more flexibility.
Before using a custom certificate, first it must be mounted with a volume. (e.g. with the entry under volumes ` - /etc/ssl/wikijs:/path/to/folder:ro`).
Unfortunately this is not fully backwards compatible, because the `SSL_PROVIDER` env variable has been added, so that all users currently using LE need to set this to `letsencrypt`.
Maybe letsencrypt could be used as default so that old environment variables still work.
I would update the wiki accordingly if this PR is considered.