Add width and height to allowed markdown attributes
Currently if using the markdown files outside of wiki.js (In my case with pandoc) it's impossible to specify the size of an image that works in both places as the width/height attributes are stripped in wiki.js. Whilst it may be possible to do this using a class (Which is allowed), this does not allow for specific per image width/height values. 

Is there any reason that these cannot be included in the allowed attributes?