fix: inline math interpreted as attributes
When using inline math ($e^{-x^2}$) the curly braces are interpreted as
attributes by markdown-it-attrs. Since most of the times they are not
valid attributes they simply get removed.

This patch escapes the curly braces (the default attribute delimiter),
fixing the KaTeX rendering errors.

It would be nice to simply skip that rule for `katex_inline` block types
but as far as I know markdown-it-attrs doesn't have such an option.

Fixes #1581