fix: loginRedirect doesn't work for non local strategies
For the local authentication strategy the loginRedirect cookie is read in login.vue and will redirect the user to the page they requested before needing to login. If you aren't using the local authentication strategy, but instead are using a different strategy, like OpenID Connect etc then logins aren't redirected back to the desired location. Fix this so behavior is the same whether using the local strategy or a different one. Fix based on what is done in login.vue.