feat: add newline to the end of page content
When pages are stored in a bidirectional storage (e.g. git), it's expected
that people can edit and update the files there. Many text editor add a
newline or warn about missing newlines at the end of files.
Adding tailing newlines when storing pages to disk, will avoid this issues.