Fix existing auth strategies not getting new props
This pull request should fix an existing authentication strategy not getting new properties after an update. It does this by using the existing function to pull in new properties and their defaults, and if any new properties are detected, write the strategy back to the database with the new defaults.

Something of note, a bad auth strategy update that includes breaking default properties can break an installation, so people need to keep this in mind if this PR is merged.